--- engine/src/main/resources/assets/ui/menu/moduleTypes.ui
@@ -1,182 +1,0 @@
-{
-  "type": "advanceModuleFilter",
-  "skin": "engine:mainMenu",
-  "contents": {
-    "type": "relativeLayout",
-    "contents": [
-      {
-        "type": "UIBox",
-        "id": "container",
-        "layoutInfo": {
-          "width": 500,
-          "use-content-height": true,
-          "position-horizontal-center": {},
-          "position-vertical-center": {}
-        },
-        "content": {
-          "type": "ColumnLayout",
-          "columns": 1,
-          "verticalSpacing": 8,
-          "contents": [
-            {
-              "type": "UILabel",
-              "text": "Advanced Filter",
-              "family": "title"
-            },
-            {
-              "type": "UISpace",
-              "size": [
-                1,
-                16
-              ]
-            },
-            {
-              "type": "RowLayout",
-              "contents": [
-                {
-                  "type": "UICheckbox",
-                  "id":"libraryCheckbox",
-                  "family": "option-grid",
-                  "layoutInfo": {
-                    "relativeWidth": 0.04
-                  }
-                },
-                {
-                  "type": "UILabel",
-                  "id": "library",
-                  "text": "Library",
-                  "enabled": true,
-                  "layoutInfo": {
-                    "relativeWidth": 0.15
-                  }
-                }
-              ]
-            },
-            {
-              "type": "RowLayout",
-              "contents": [
-                {
-                  "type": "UICheckbox",
-                  "id": "assetCheckbox",
-                  "family": "option-grid",
-                  "layoutInfo": {
-                    "relativeWidth": 0.04
-                  }
-                },
-                {
-                  "type": "UILabel",
-                  "text": "Asset ",
-                  "id": "asset",
-                  "layoutInfo": {
-                    "relativeWidth": 0.15
-                  }
-                }
-              ]
-            },
-            {
-              "type": "RowLayout",
-              "contents": [
-                {
-                  "type": "UICheckbox",
-                  "id": "worldCheckbox",
-                  "family": "option-grid",
-                  "layoutInfo": {
-                    "relativeWidth": 0.04
-                  }
-                },
-                {
-                  "type": "UILabel",
-                  "text": "World ",
-                  "id": "world",
-                  "layoutInfo": {
-                    "relativeWidth": 0.15
-                  }
-                }
-              ]
-            },
-            {
-              "type": "RowLayout",
-              "contents": [
-                {
-                  "type": "UICheckbox",
-                  "id": "specialCheckbox",
-                  "family": "option-grid",
-                  "layoutInfo": {
-                    "relativeWidth": 0.04
-                  }
-                },
-                {
-                  "type": "UILabel",
-                  "text": "Special",
-                  "id": "special",
-                  "id": "port",
-                  "layoutInfo": {
-                    "relativeWidth": 0.15
-                  }
-                }
-              ]
-            },
-            {
-              "type": "RowLayout",
-              "contents": [
-                {
-                  "type": "UICheckbox",
-                  "id": "augmentationCheckbox",
-                  "family": "option-grid",
-                  "layoutInfo": {
-                    "relativeWidth": 0.04
-                  }
-                },
-                {
-                  "type": "UILabel",
-                  "text": "Augmentation",
-                  "id": "augmentation",
-                  "id": "augmentation",
-                  "layoutInfo": {
-                    "relativeWidth": 0.26
-                  }
-                }
-              ]
-            },
-            {
-              "type": "RowLayout",
-              "contents": [
-                {
-                  "type": "UICheckbox",
-                  "id": "gameplayCheckbox",
-                  "family": "option-grid",
-                  "layoutInfo": {
-                    "relativeWidth": 0.04
-                  }
-                },
-                {
-                  "type": "UILabel",
-                  "text": "Gameplay",
-                  "id": "gameplay",
-                  "layoutInfo": {
-                    "relativeWidth": 0.20
-                  }
-                }
-              ]
-            }
-          ]
-        }
-      },
-      {
-        "type": "UIButton",
-        "text": "Back",
-        "id": "back",
-        "layoutInfo": {
-          "width": 100,
-          "height": 25,
-          "position-horizontal-center": {},
-          "position-bottom":{
-            "widget":"container",
-            "target":"BOTTOM",
-            "offset": -50
-          }
-        }
-      }
-    ]
-  }
-}
